refactor: move pending reservation type
